finish error handling

+++ services/categoryService.js
@@ -1,18 +1,19 @@
 const CategoryModel = require('../models/categoryModel');
 const slugify = require('slugify');
 const {asyncHandler}=require('../middlewares/asyncHandler');
-
+const globalErrorHandler = require('../utils/apiError');
+const mongoose = require('mongoose')
 
 // @desc get list of categories
 // @route get /api/categories
 // @access public
-exports.getAllCategories=asyncHandler(async(req,res)=>{
+exports.getAllCategories=asyncHandler(async(req,res,next)=>{
         const limit = +req.query.limit||2;
         const page = +req.query.page||1;
         const skip = limit*page-limit;
         const categories = await CategoryModel.find().limit(limit).skip(skip);
         if(!categories)
-                return res.status(404).json({status:"FAIL",data:{categories:"there is no categories yet"}});
+                return next(new globalErrorHandler("there is no categories yet",404));
         res.status(200).json({status:"SUCCESS",page:page,data:{categories}});
 });
 
@@ -20,11 +21,14 @@ exports.getAllCategories=asyncHandler(async(req,res)=>{
 // @desc get specific category
 // @route get /api/categories/:id
 // @access public
-exports.getCategory=asyncHandler(async(req,res)=>{
-        const id = +req.query.id;
+exports.getCategory=asyncHandler(async(req,res,next)=>{
+        const id = req.params.id;
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+                return next(new globalErrorHandler("Invalid ID format", 400));
+        }
         const category = await CategoryModel.findById(id);
         if(!category)
-                return res.status(404).json({status:"FAIL",msg:"category not found"});
+                return next(new globalErrorHandler("category not found",404));
         res.status(200).json({status:"SUCCESS",data:{category}});
 });
 
@@ -42,22 +46,31 @@ exports.createCategory= asyncHandler(async(req,res)=>{
 // @desc update specific category
 // @route put /api/categories/:id
 // @access private
-exports.UpdateCategory=asyncHandler(async(req,res)=>{
-        const id = +req.query.id;
+exports.UpdateCategory=asyncHandler(async(req,res,next)=>{
+        const id = req.params.id;
         const name=req.body.name;
+
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+                return next(new globalErrorHandler("Invalid ID format", 400));
+        }
+
         const newCategory = await CategoryModel.findByIdAndUpdate(id,{name,slug:slugify(name)},{new:true});
         if(!newCategory)
-                return res.status(404).json({status:"FAIL",msg:"category not found"});
+                return next(new globalErrorHandler("category not found",404));
         res.status(200).json({status:"SUCCESS",data:{newCategory}});
 });
 
 // @desc delete category by id
 // @route delete /api/categories
 // @access private 
-exports.deleteCategory=asyncHandler(async(req,res)=>{
-        const id = +req.query.id;
+exports.deleteCategory=asyncHandler(async(req,res,next)=>{
+        const id = req.params.id;
+
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+                return next(new globalErrorHandler("Invalid ID format", 400));
+        }
         const deletedCategory = await CategoryModel.findByIdAndDelete(id);
         if(!deletedCategory)
-                return res.status(404).json({status:"FAIL",data:{categories:"category not found"}});
+                return next(new globalErrorHandler("category not found",404));
         res.status(200).json({status:"SUCCESS",msg:"deleted"});
 });
